LPS-124047 Renaming return type and moving resolver to the item-selector module
