+++ modules/apps/document-library/document-library-item-selector-web/src/main/java/com/liferay/document/library/item/selector/web/internal/video/DLVideoItemSelectorView.java
@@ -22,7 +22,7 @@ import com.liferay.item.selector.ItemSelectorReturnType;
 import com.liferay.item.selector.ItemSelectorView;
 import com.liferay.item.selector.criteria.FileEntryItemSelectorReturnType;
 import com.liferay.item.selector.criteria.URLItemSelectorReturnType;
-import com.liferay.item.selector.criteria.VideoURLItemSelectorReturnType;
+import com.liferay.item.selector.criteria.VideoEmbeddableHTMLItemSelectorReturnType;
 import com.liferay.item.selector.criteria.video.criterion.VideoItemSelectorCriterion;
 import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.ListUtil;
@@ -82,6 +82,6 @@ public class DLVideoItemSelectorView
 			ListUtil.fromArray(
 				new FileEntryItemSelectorReturnType(),
 				new URLItemSelectorReturnType(),
-				new VideoURLItemSelectorReturnType()));
+				new VideoEmbeddableHTMLItemSelectorReturnType()));
 
 }
